requirement.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
pkg-resources==0.0.0
PyJWT==1.7.1
requests==2.24.0
urllib3==1.25.9
Werkzeug==1.0.1
